requirements.txt
streamlit
requests
python-dotenv
Pillow
python-magic
